Fix bug causing jobs to be endless and duplicated
