+++ src/containers/BackgroundTasks/types/CheckOrderInvoicesStatus.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { JobStatusEnum } from "./../../types/globalTypes";
+import { JobStatusEnum } from "./../../../types/globalTypes";
 
 // ====================================================
 // GraphQL query operation: CheckOrderInvoicesStatus
