+++ src/products/queries.ts
@@ -300,12 +300,3 @@ export const useCreateMultipleVariantsData = makeQuery<
   CreateMultipleVariantsData,
   CreateMultipleVariantsDataVariables
 >(createMultipleVariantsData);
-
-export const checkExportFileStatus = gql`
-  query CheckExportFileStatus($id: ID!) {
-    exportFile(id: $id) {
-      id
-      status
-    }
-  }
-`;
